refactor run monitor stuff into separate file for sanity wrap watch stream with recovery from conncetion reset by peer
